backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1
celery==5.3.4
openai==1.3.7
boto3==1.34.0
azure-identity==1.15.0
azure-mgmt-compute==30.4.0
azure-mgmt-network==25.2.0
azure-mgmt-resource==23.1.0
psycopg2-binary==2.9.9
email-validator==2.1.0
Jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0

prometheus-client==0.19.0
psutil==5.9.6